Two New Traits. (#1792)
Plateled Factories but for IPCs, at the cost of 10 trait points, heals
0,8.

Bionic legs roundstart, at the cost of 6 points, since they are not as
useful as the arms.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Platelet Factories for IPC feel good since it made, in my opinion, and
in the opinion of many others, no sense that IPCs had nothing for regen,
as for bionic legs, consider it my personal mald PR for not having
bionic legs from the start.
-->
Plateled Factories but for IPCs, at the cost of 10 trait points, heals
0,8.

Bionic legs roundstart, at the cost of 6 points, since they are not as
useful as the arms.

Platelet Factories for IPC feel good since it made, in my opinion, and
in the opinion of many others, no sense that IPCs had nothing for regen,
as for bionic legs, consider it my personal mald PR for not having
bionic legs from the start.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- add: A trait for IPCs, which is essentially a platelet factories, at
the same trait cost, etc, as the normal platelet factories, and bionic
legs, which are just speed legs roundstart.

(cherry picked from commit f6d0296d45568c1006b2dea7fc927f588e9db0f1)