Two New .25 Caseless Guns (#1794)
# Description

This PR adds a new integrally suppressed submachine gun to the game,
which is called the FPA-90. It is notably the first automatic weapon to
fire the .25 caseless ammo type, which it does so in extremely quick 3
round bursts with a moderate cooldown between bursts. This item can be
either obtained from the Syndicate Uplink where it's intended primarily
for "Stealth Ops" nukies, or it can be manufactured after researching
Advanced Weapons in Arsenal T3, for a fairly sizeable amount of
aluminium and steel.

Since Security could not previously print .25 caseless, this PR also
adds the ability to print caseless rifle magazines. Also included for
the uplink are a variety of magazines for the FPA-90(and other weapons).

For the crew side of things, there is also the bullpup caseless
integrally supressed assault rifle, called the BRDI-R25. Unlike the
FPA-90, it requires wielding to be used effectively. It's currently
obtained via Prototype Weapons in arsenal t3, and then manufactured at a
security techfab. The BRDI has two radically different firing modes, a 5
round burst with extreme rate of fire and poor accuracy, and an auto
mode that has much lower firing rate and more manageable recoil.
Essentially, we have a "Fuller-Auto" gun.

There is also a unique antique BRDI-R25 available to the Head of
Security via his special weapon loadout picks.

<details><summary><h1>Media</h1></summary>
<p>
Available in research

![image](https://github.com/user-attachments/assets/5b28f1f8-f954-4eea-895f-c18f0905a2ac)

And in the uplink catalog

![image](https://github.com/user-attachments/assets/b5e1cd96-5592-462f-85a4-436ecd5744d2)

![image](https://github.com/user-attachments/assets/a92691b7-f1f4-44b7-9782-462d284d0932)

In the techfab(after researching it)

![image](https://github.com/user-attachments/assets/a6d56942-2987-4a59-8e04-b069edbb31b1)

Firing it:

https://github.com/user-attachments/assets/fd82df27-bf26-4351-be9d-c4da71d8a0e2

</p>
</details>

# Changelog

:cl:
- add: Added the FPA-90, an integrally supressed submachine gun. It can
be found either in the syndicate uplink, or manufactured at a security
techfab after Epistemics unlocks Advanced Weapons.
- add: Added the BRDI-R25, an integrally supressed bullpup rifle
chambered in .25 Caseless. It has an "Extremely fast" burst fire mode,
and a "More controllable" full-auto/semi mode. Unlike the FPA-90, it is
not available to syndicate agents. It can be manufactured after
researching Prototype Weapons.
- add: Antique variant of the BRDI-R25, which can be taken by the Head
of Security as his personal weapon.
- add: Added a variety of different ammunition options to Syndicate
uplinks.
- add: Security Techfabs can now manufacture .25 caseless rifle ammo and
magazines.

(cherry picked from commit 6647eca1041b550ab4569b434e80cd0b7c20099a)