Update pebble.yml (#1826)
# Description

Fixes a small issue with pebble that was causing test fails. Also
removes the captain's laser from being mapped. He gets it from loadouts
now, and chooses between it and other options that take the place of the
traitor steal objective target. It should not be mapped like this
anymore because that would create a duplicate of an item which is only
supposed to be one of on the map.

(cherry picked from commit 778ee231b1cfb62ec3ca48a80fe9761eb9fb3781)