Minor Fixes to My Two Traits PR. (#1833)
Fixed the nano bot trait having a bit too much healing and actually got
the bionic leg trait to work.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Nano bot trait nerf and Bionic Leg fix.

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Could be defined as incompetence and also a "this is the first time I do
this" kind of error, regardless, platelet factories for IPC now only
heals 0,6 instead of 0,9 and Bionic Legs ACTUALLY do work this time
around..

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- tweak: Tweaked the nano bot trait value from 0,9 to 0,6.
- fix: Bionic legs DO work now, yay.

(cherry picked from commit c03b719d931b67ba4b01eea1dfd2d77b840047bd)